multiDocsQAwithLangchain/requirements.txt
python-dotenv
google-generativeai
pyPDF2
langchain
chromadb
langchain_google_genai
streamlit
faiss-cpu
langchain-community
